% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateAmbiguity.R
\name{CalculateAmbiguity}
\alias{CalculateAmbiguity}
\title{Calculation of the ambiguity for triangular and trapezoidal fuzzy numbers}
\usage{
CalculateAmbiguity(fuzzyNumber, increases = FALSE)
}
\arguments{
\item{fuzzyNumber}{Input data consist of triangular or trapezoidal fuzzy numbers.}

\item{increases}{If \code{TRUE} is used, then the initial data should consist of the fuzzy numbers in the form:
left increment of the support, left end of the core, right end of the core,
right increment of the support. Otherwise, the default value \code{FALSE} is used and the fuzzy numbers should be given in the form:
left end of the support, left end of the core, right end of the core,
right end of the support.}
}
\value{
This function returns vector of double values.
Each output value is equal to the ambiguity of the respective fuzzy number.
}
\description{
\code{CalculateAmbiguity} returns the ambiguity of the triangular or trapezoidal fuzzy number (see, e.g., (Ban et al., 2015),
(Grzegorzewski and Romaniuk, 2022)).
}
\details{
The input data should consist of triangular or trapezoidal fuzzy numbers, given as a single vector or a whole matrix.
In each row, there should be a single fuzzy number in one of the forms:
\enumerate{
\item left end of the support, left end of the core, right end of the core, right end of the support, or
\item left increment of the support, left end of the core, right end of the core, right increment of the support.
}
In this second case, the parameter \code{increases=TRUE} has to be used.

Then for each fuzzy number, its characteristics, known as the ambiguity of fuzzy number, is calculated.
For the respective formulas, see, e.g., (Ban et al., 2015), (Grzegorzewski and Romaniuk, 2022).
}
\examples{

# prepare some fuzzy numbers (first type of the initial sample)

fuzzyValues <- matrix(c(0.25,0.5,1,1.25,0.75,1,1.5,2.2,-1,0,0,2),
ncol = 4,byrow = TRUE)

# calculate the ambiguity of the first fuzzy number

CalculateAmbiguity(fuzzyValues[1,])

# calculate the ambiguity for the whole matrix

CalculateAmbiguity(fuzzyValues)

# prepare some fuzzy numbers (second type of the initial sample)

fuzzyValuesInc <- matrix(c(0.25,0.5,1,0.25,0.25,1,1.5,0.7,1,0,0,2),
ncol = 4,byrow = TRUE)

# calculate the ambiguity of the first fuzzy number

CalculateAmbiguity(fuzzyValuesInc[1,], increases = TRUE)


}
\references{
Ban, A.I., Coroianu, L., Grzegorzewski, P. (2015)
Fuzzy Numbers: Approximations, Ranking and Applications
Institute of Computer Sciences, Polish Academy of Sciences

Grzegorzewski, P., Romaniuk, M. (2022)
Bootstrap methods for fuzzy data
Uncertainty and Imprecision in Decision Making and Decision Support: New Advances, Challenges, and Perspectives, pp. 28-47
Springer
}
\seealso{
\code{\link{CalculateFuzziness}} for calculation of the fuzziness,
\code{\link{CalculateValue}} for calculation of the value,
\code{\link{CalculateAmbiguityR}} for calculation of the right-hand ambiguity,
\code{\link{CalculateAmbiguityL}} for calculation of the left-hand ambiguity,
\code{\link{CalculateExpValue}} for calculation of the expected value,
\code{\link{CalculateWidth}} for calculation of the width

Other characteristics of fuzzy numbers functions: 
\code{\link{CalculateAmbiguityL}()},
\code{\link{CalculateAmbiguityR}()},
\code{\link{CalculateExpValue}()},
\code{\link{CalculateFuzziness}()},
\code{\link{CalculateValue}()},
\code{\link{CalculateWidth}()}
}
\concept{characteristics of fuzzy numbers functions}
