% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamplingMethods.R
\docType{data}
\name{resamplingMethods}
\alias{resamplingMethods}
\title{A vector containing names of all resampling methods.}
\format{
An object of class \code{character} of length 6.
}
\usage{
resamplingMethods
}
\description{
\code{resamplingMethods} is a vector containing names of all resampling methods.
}
\keyword{datasets}
