% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamplingMethods.R
\docType{data}
\name{resamplingMethods}
\alias{resamplingMethods}
\title{A vector containing names of all resampling methods.}
\format{
An object of class \code{character} of length 7.
}
\usage{
resamplingMethods
}
\value{
This function returns a vector of strings.
}
\description{
\code{resamplingMethods} is a vector containing names of all resampling methods.
}
\examples{

# check the names of the available resampling methods

resamplingMethods

}
\keyword{datasets}
